Allow buying decimal amount of stocks

+++ Mods/SOD.StockMarket/Implementation/Cruncher/Content/AppInstantSellInterface.cs
@@ -32,7 +32,7 @@ namespace SOD.StockMarket.Implementation.Cruncher.Content
             MapButton("+10", () => IncreaseAmount(10));
             MapButton("+100", () => IncreaseAmount(100));
             MapButton("+1000", () => IncreaseAmount(1000));
-            MapButton("Max", () => IncreaseAmount(int.MaxValue));
+            MapButton("Max", () => IncreaseAmount(decimal.MaxValue));
             MapButton("Reset", ResetAmount);
             MapButton("Sell", InstantSell);
 
@@ -42,9 +42,9 @@ namespace SOD.StockMarket.Implementation.Cruncher.Content
             Lib.Time.OnMinuteChanged += UpdateInfo;
         }
 
-        private int _currentAmount = 0;
+        private decimal _currentAmount = 0;
         private decimal _currentPayment = 0;
-        private void IncreaseAmount(int amount)
+        private void IncreaseAmount(decimal amount)
         {
             BalanceAmountAndCostBasedOnCurrentPrices(amount);
 
@@ -72,12 +72,12 @@ namespace SOD.StockMarket.Implementation.Cruncher.Content
             ResetAmount();
         }
 
-        private void BalanceAmountAndCostBasedOnCurrentPrices(int amount)
+        private void BalanceAmountAndCostBasedOnCurrentPrices(decimal amount)
         {
             var tradeController = Plugin.Instance.Market.TradeController;
             var maxAffordableAmount = tradeController.GetInvestedVolume(_stock);
 
-            var newAmount = amount == int.MaxValue ? maxAffordableAmount : (_currentAmount + amount);
+            var newAmount = amount == decimal.MaxValue ? maxAffordableAmount : (_currentAmount + amount);
             if (newAmount > maxAffordableAmount)
                 newAmount = maxAffordableAmount;
