+++ Mods/SOD.StockMarket/Implementation/Trade/TradeOrder.cs
@@ -4,13 +4,13 @@
     {
         public int StockId { get; set; }
         public decimal Price { get; set; }
-        public int Amount { get; set; }
+        public decimal Amount { get; set; }
         public OrderType OrderType { get; set; }
         public bool Completed { get; set; }
 
         public TradeOrder() { }
 
-        internal TradeOrder(OrderType orderType, int stockId, decimal price, int amount)
+        internal TradeOrder(OrderType orderType, int stockId, decimal price, decimal amount)
         {
             StockId = stockId;
             Price = price;
